Allow backends to declare their own auto-activation conditions
Previously, all the logic for auto-selecting backends was housed
inside the "amx-completing-read-auto" function, which needed to be
modified every time a new backend was added. Now the condition for
automatically using a backend is declared as part of the backend, and
"amx-completing-read-auto" just uses these declarations to select the
appropriate backend.