Don't require restored commands to be defined in the current Emacs
Since amx generally restores its history during init, it's very likely
that some commands will not be defined yet, e.g. because the package
that defines them is loaded later in the init file or is autoloaded.
Hence, when restoring history from a file, amx needs to restore all
commands mentioned in the history file regardless of whether they are
defined in the current Emacs.

This should address #48.

A test is added for this behavior. (This test fails without the change
to amx.el.)